# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Point of Sale API",
        "description": "API Documentation for Point of Sale Application",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/v1/stock-products": {
            "get": {
                "tags": [
                    "Stock Products"
                ],
                "summary": "List all stock products with optional filters",
                "operationId": "77a7ccc791351c74c4b50d7c3930f587",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Filter by product ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filter by user ID (branch)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of stock products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StockProduct"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Stock Products"
                ],
                "summary": "Create a new stock product",
                "operationId": "8354a16278b4ff8339fe3972031ca4d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StockProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Stock product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StockProduct"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/stock-products/{id}": {
            "get": {
                "tags": [
                    "Stock Products"
                ],
                "summary": "Get a specific stock product",
                "operationId": "8220b33705e086dcc54c95b79adb99ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stock product details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StockProduct"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Stock product not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Stock Products"
                ],
                "summary": "Update a stock product",
                "operationId": "ba9953a8d917c0f9202e27a18623cc11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StockProduct"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StockProduct"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Stock product not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Stock Products"
                ],
                "summary": "Delete a stock product",
                "operationId": "feccc6de5f7e60bfe49b6b1327ce8af5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Stock product deleted successfully"
                    },
                    "404": {
                        "description": "Stock product not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StockProductRequest": {
                "required": [
                    "product_id",
                    "user_id",
                    "quantity"
                ],
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "minimum": 0,
                        "example": 100
                    },
                    "note": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "image": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "StockProduct": {
                "required": [
                    "product_id",
                    "user_id",
                    "quantity"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}